フレームのコンテキストメニューの実装を別のファイルに移動
